#356 - Move rest api to v4
- move all endpoints from v3 to v4
- merge TerminologyEsRestController into TerminologyRestController
- rename CodeableConceptEsRestController to CodeableConceptRestController
- rename all "skipValidation" parameters to "skip-validation"
- use path constants instead of explicit strings more consistently
- update swagger doc
