+++ .github/scripts/post-test-query.sh
@@ -10,7 +10,7 @@ access_token="$(curl -s --request POST \
   --data scope=openid | jq '.access_token' | tr -d '"')"
 
 response=$(curl -s -i \
-  --url http://localhost:8091/api/v3/query \
+  --url http://localhost:8091/api/v4/query \
   --header "Authorization: Bearer $access_token" \
   --header 'Content-Type: application/json' \
   --data '{
